Lots of cleanups, bash fixes, plus the below.
Remove attempt to delete tag when exiting before it is created.
Use git log formatting arguments instead of requesting the entire log.
Add exit signal trap to remove temporary file in case of any exit.
Move duplicate change pushing outside of branches. Previously this was not executed if the branch was master - why?